requirements.txt
absl-py==0.13.0
alembic==1.6.5
aniso8601==9.0.1
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.4
flask-accepts==0.18.2
flask-marshmallow==0.14.0
Flask-Migrate==2.1.1
flask-restplus==0.13.0
flask-restx==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
gast==0.3.3
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.38.1
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
importlib-metadata==4.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.1
mypy-extensions==0.4.3
numpy==1.18.5
oauthlib==3.1.1
opt-einsum==3.3.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.19
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
